知乎回答及专栏文章的爬虫-废弃/requirements.txt
requests
bs4
lxml
pillow
html2text